+++ lib/panda-editor.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require "panda/editor/version"
+require "panda/editor/engine"
+require "panda/editor"